renamed _opt.py to opt.py
finished BEOPT attrs refactoring
