+++ mypy.ini
@@ -6,7 +6,7 @@
 
 # explicitly blacklist files, this means we can easily add in stricter type checks
 # by removing them from the blacklist
-[mypy-quemb.molbe.ube,quemb.molbe.be_parallel,quemb.molbe.autofrag,quemb.molbe.eri_onthefly,quemb.molbe.fragment,quemb.molbe.helper,quemb.molbe.lchain,quemb.molbe.lo,quemb.molbe.mbe,quemb.molbe.misc,quemb.molbe._opt,quemb.molbe.pfrag,quemb.molbe.solver]
+[mypy-quemb.molbe.ube,quemb.molbe.be_parallel,quemb.molbe.autofrag,quemb.molbe.eri_onthefly,quemb.molbe.fragment,quemb.molbe.helper,quemb.molbe.lchain,quemb.molbe.lo,quemb.molbe.mbe,quemb.molbe.misc,quemb.molbe.opt,quemb.molbe.pfrag,quemb.molbe.solver]
     disallow_untyped_defs = False
     check_untyped_defs = False
