+++ src/quemb/kbe/pbe.py
@@ -14,9 +14,9 @@ from pyscf.pbc.df.df_jk import _ewald_exxdiv_for_G0
 from quemb.kbe.lo import Mixin_k_Localize
 from quemb.kbe.misc import print_energy, storePBE
 from quemb.kbe.pfrag import Frags
-from quemb.molbe._opt import BEOPT
 from quemb.molbe.be_parallel import be_func_parallel
 from quemb.molbe.helper import get_eri, get_scfObj, get_veff
+from quemb.molbe.opt import BEOPT
 from quemb.molbe.solver import be_func
 from quemb.shared.config import settings
 from quemb.shared.external.optqn import (
