+++ src/quemb/molbe/opt.py
@@ -2,12 +2,18 @@
 
 
 import numpy
+from attrs import define
+from numpy import float64
 
 from quemb.molbe.be_parallel import be_func_parallel
+from quemb.molbe.fragment import fragpart
 from quemb.molbe.solver import be_func
 from quemb.shared.external.optqn import FrankQN
+from quemb.shared.manage_scratch import WorkDir
+from quemb.shared.typing import KwargDict, Matrix
 
 
+@define
 class BEOPT:
     """Perform BE optimization.
 
@@ -18,84 +24,65 @@ class BEOPT:
 
     Parameters
     ----------
-    pot : list
+    pot :
        List of initial BE potentials. The last element is for the global
        chemical potential.
     Fobjs : quemb.molbe.fragment.fragpart
        Fragment object
-    Nocc : int
+    Nocc :
        No. of occupied orbitals for the full system.
-    enuc : float
+    enuc :
        Nuclear component of the energy.
-    solver : str
+    solver :
        High-level solver in bootstrap embedding. 'MP2', 'CCSD', 'FCI' are supported.
        Selected CI versions,
        'HCI', 'SHCI', & 'SCI' are also supported. Defaults to 'MP2'
-    only_chem : bool
+    only_chem :
        Whether to perform chemical potential optimization only.
        Refer to bootstrap embedding literatures.
-    nproc : int
+    nproc :
        Total number of processors assigned for the optimization. Defaults to 1.
        When nproc > 1, Python multithreading
        is invoked.
-    ompnum : int
+    ompnum :
        If nproc > 1, ompnum sets the number of cores for OpenMP parallelization.
        Defaults to 4
-    max_space : int
+    max_space :
        Maximum number of bootstrap optimizaiton steps, after which the optimization
        is called converged.
-    conv_tol : float
+    conv_tol :
        Convergence criteria for optimization. Defaults to 1e-6
-    ebe_hf : float
+    ebe_hf :
        Hartree-Fock energy. Defaults to 0.0
     """
 
-    def __init__(
-        self,
-        pot,
-        Fobjs,
-        Nocc,
-        enuc,
-        solver="MP2",
-        nproc=1,
-        ompnum=4,
-        only_chem=False,
-        hf_veff=None,
-        hci_pt=False,
-        hci_cutoff=0.001,
-        ci_coeff_cutoff=None,
-        select_cutoff=None,
-        max_space=500,
-        conv_tol=1.0e-6,
-        relax_density=False,
-        ebe_hf=0.0,
-        scratch_dir=None,
-        **solver_kwargs,
-    ):
-        # Initialize instance attributes
-        self.ebe_hf = ebe_hf
-        self.hf_veff = hf_veff
-        self.pot = pot
-        self.Fobjs = Fobjs
-        self.Nocc = Nocc
-        self.enuc = enuc
-        self.solver = solver
-        self.iter = 0
-        self.err = 0.0
-        self.Ebe = 0.0
-        self.max_space = max_space
-        self.nproc = nproc
-        self.ompnum = ompnum
-        self.only_chem = only_chem
-        self.conv_tol = conv_tol
-        self.relax_density = relax_density
-        # HCI parameters
-        self.hci_cutoff = hci_cutoff
-        self.ci_coeff_cutoff = ci_coeff_cutoff
-        self.select_cutoff = select_cutoff
-        self.hci_pt = hci_pt
-        self.solver_kwargs = solver_kwargs
-        self.scratch_dir = scratch_dir
+    pot: list[float]
+    Fobjs: fragpart
+    Nocc: int
+    enuc: float
+    scratch_dir: WorkDir
+    solver: str = "MP2"
+    nproc: int = 1
+    ompnum: int = 4
+    only_chem: bool = False
+    hf_veff: Matrix[float64] | None = None
+
+    max_space: int = 500
+    conv_tol: float = 1.0e-6
+    relax_density: bool = False
+    ebe_hf: float = 0.0
+
+    iter: int = 0
+    err: float = 0.0
+    Ebe: float = 0.0
+
+    solver_kwargs: KwargDict | None = None
+
+    # HCI parameters
+    hci_cutoff: float = 0.001
+    ci_coeff_cutoff: float | None = None
+    select_cutoff: float | None = None
+    hci_pt: bool = False
 
     def objfunc(self, xk):
         """
